Bug 1792920, part 2 - Fix the type of the refcount argument for NS_LogAddRef and NS_LogRelease in Rust. r=xpcom-reviewers,nika
The type of the refcount argument to these functions is the C++ type
nsrefcnt, which is actually uintptr_t. This is different from the
definition of nsrefcnt in XPIDL which is currently being used. These
functions are not actually defined in XPIDL, and also uintptr_t
can't be expressed in XPIDL, so let's just change it to the type it
really is in Rust, usize. The goal of the patch is to eliminate
some uses of nsrefcnt from Rust.

Differential Revision: https://phabricator.services.mozilla.com/D159321

+++ xpcom/rust/xpcom/xpcom_macros/src/lib.rs
@@ -714,7 +714,7 @@ fn xpcom_impl(args: AttributeArgs, template: ItemStruct) -> Result<TokenStream, 
                 let new = self.__refcnt.inc();
                 ::xpcom::trace_refcnt::NS_LogAddRef(
                     self as *const _ as *mut ::xpcom::reexports::libc::c_void,
-                    new,
+                    new as usize,
                     #name_for_logs,
                     #size_for_logs,
                 );
@@ -726,7 +726,7 @@ fn xpcom_impl(args: AttributeArgs, template: ItemStruct) -> Result<TokenStream, 
                 let new = self.__refcnt.dec();
                 ::xpcom::trace_refcnt::NS_LogRelease(
                     self as *const _ as *mut ::xpcom::reexports::libc::c_void,
-                    new,
+                    new as usize,
                     #name_for_logs,
                     #size_for_logs,
                 );
